+++ pages/api/country.ts
@@ -1,7 +1,11 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { Prisma, Country } from '@prisma/client';
+import * as yup from 'yup';
+import { unstable_getServerSession } from 'next-auth';
 import prisma from '../../lib/prisma';
 import { ApiResponse } from '../../lib/types/api';
+import { deleteSchema } from '../../utils/api/validation';
+import { authOptions } from './auth/[...nextauth]';
 
 export default async function handler(
   req: NextApiRequest,
@@ -12,9 +16,44 @@ export default async function handler(
   >
 ) {
   try {
+    const session = await unstable_getServerSession(
+      req as NextApiRequest,
+      res as NextApiResponse,
+      authOptions(req as NextApiRequest, res as NextApiResponse)
+    );
     const { method } = req;
     if (method === 'POST') {
+      if (!session || session.user.role !== 'ADMIN') {
+        res.status(401).json({
+          name: 'UNAUTHORIZED',
+          success: false,
+          message: 'Unauthorized',
+        });
+        return;
+      }
       const { input } = req.body;
+      const postSchema = yup.object({
+        input: yup
+          .object({
+            data: yup
+              .object({
+                name: yup
+                  .string()
+                  .required()
+                  .matches(/^[A-Za-z]+$/),
+                translation: yup
+                  .string()
+                  .required()
+                  .matches(/^[\u0590-\u05fe]+$/),
+                lockTranslation: yup.boolean().required(),
+                show: yup.boolean(),
+              })
+              .required(),
+            include: yup.object(),
+          })
+          .required(),
+      });
+      await postSchema.validate({ input });
       const country = await prisma.country.create({
         ...input,
       });
@@ -31,13 +70,51 @@ export default async function handler(
       });
       res.status(200).json({ success: true, data: { countries } });
     } else if (method === 'PUT') {
+      if (!session || session.user.role !== 'ADMIN') {
+        res.status(401).json({
+          name: 'UNAUTHORIZED',
+          success: false,
+          message: 'Unauthorized',
+        });
+        return;
+      }
       const { input } = req.body;
+      const putSchema = yup.object({
+        input: yup
+          .object({
+            where: yup
+              .object({
+                id: yup.string().required(),
+              })
+              .required(),
+            data: yup
+              .object({
+                name: yup.string().matches(/^[A-Za-z]+$/),
+                translation: yup.string().matches(/^[\u0590-\u05fe]+$/),
+                lockTranslation: yup.boolean(),
+                show: yup.boolean(),
+              })
+              .required(),
+            include: yup.object(),
+          })
+          .required(),
+      });
+      await putSchema.validate({ input });
       const update: Country = await prisma.country.update({
         ...input,
       });
       res.status(200).json({ success: true, data: update });
     } else if (method === 'DELETE') {
+      if (!session || session.user.role !== 'ADMIN') {
+        res.status(401).json({
+          name: 'UNAUTHORIZED',
+          success: false,
+          message: 'Unauthorized',
+        });
+        return;
+      }
       const { action, input } = req.body;
+      await deleteSchema.validate({ action, input });
       if (action === 'deleteMany') {
         const deleteMany = await prisma.country.deleteMany({
           ...input,
